DEV-1728 PR #276 review round 1: Sonar gate, CodeRabbit, Codex
Sonar quality gate (new_duplicated_lines_density 10.6% vs 3% threshold) —
extract the cross-model test fixtures + SQL-shape helpers into
`tests/_cross_model_chain.py` (underscore-prefixed so pytest skips collection,
same convention as `tests/_engine_helpers.py`). The `orders_x → customers_v2 →
regions → countries` chain builders and the `_norm` / `_extract_cte_body` /
`_split_at_ranked_subquery` / `_joinback_on_predicate` helpers were copied
per-module; the copies were byte-identical, so the duplication bought nothing.
test_dev1728, test_dev1708 and test_sql_generator now import the one definition
(123 duplicated lines removed from the new file alone). `_customers_v2` carries
the union of the derived-column matrix both suites exercise, and each builder
keeps its `extra_columns` hook for module-specific shapes.

Sonar python:S1871 (cross_model_planner.py) — after DEV-1728 the `ColumnKey`
and `ColumnSqlKey` branches of `_compute_shared_grain_slots` were byte-identical.
Merge them: resolve the path per key kind, then apply one shared prefix test.
The NOSONAR(S3776) suppression goes with them — it described "three branches"
that no longer exist and the function is now well under the complexity
threshold.

CodeRabbit (tests/test_dev1728_derived_shared_grain.py) —
`test_where_filter_on_derived_grain_dim` asserted the grain expression and the
dotted alias but never that the filter itself survived, so it would have passed
had the generator dropped the predicate. Assert the rendered
`regions.population > 50`.

CodeRabbit nitpick — imports at the top: the `sqlglot` import moved to module
scope with the shared helper, and the five function-local
`from slayer.core.models import ModelJoin` in the DuckDB integration file are
hoisted to the existing module-level import (four were pre-existing; leaving
them while fixing one would be inconsistent).

Codex — `_val_<n>` collision avoidance reserved only the SEMANTIC column name,
but a `<relation>.*` projection exports the PHYSICAL name, and the two differ
whenever a column renames its source (`Column(name="renamed", sql="_val_0")`).
Both sites now go through `SQLGenerator._reserve_model_column_names`, which
reserves the semantic name plus the physical name when `Column.sql` is a bare
identifier. This also closes the same hole on the pre-existing host-path
materialisation. Pinned by a test that mints `_val_1` where the old behaviour
minted `_val_0`. Columns absent from the model remain outside what SLayer can
see without reflection — documented on the helper.

Full non-integration suite green (8489 passed); DuckDB integration 52 passed;
ruff clean.
