+++ slayer/sql/generator.py
@@ -689,6 +689,13 @@ _TRAILING_LIMIT_OFFSET_RE = re.compile(
 )
 _TRAILING_LIMIT_RE = re.compile(r"(?is)\s*LIMIT\s+\d+\s*\Z")
 
+# A ``Column.sql`` that is just an unqualified identifier — i.e. the column
+# renames a physical column rather than computing an expression. Used to
+# reserve star-exported physical names against ``_val_<n>`` collisions
+# (DEV-1728). Deliberately rejects dots: ``regions.population`` is a crossing
+# reference, not a column of the star-projected relation.
+_BARE_IDENT_RE = re.compile(r"[A-Za-z_][A-Za-z0-9_]*")
+
 
 def _strip_trailing_pagination(sql: str) -> str:
     """DEV-1444: remove trailing ORDER BY / LIMIT / OFFSET clauses that
@@ -765,6 +772,31 @@ class SQLGenerator:
         site in this module — pinned by test_dev1726_cte_case_folding — so a
         new allocation path cannot silently lose dialect awareness."""
         return AliasAllocator(folds_case=dialect_folds_case(self.dialect))
+
+    @staticmethod
+    def _reserve_model_column_names(allocator: AliasAllocator, model) -> None:
+        """Reserve every name a ``<relation>.*`` projection of ``model`` can
+        export, so a minted ``_val_<n>`` (Law-2 materialisation) never shadows a
+        real column (DEV-1728 / Codex F6).
+
+        Both the SEMANTIC name and — when ``Column.sql`` is a bare identifier —
+        the PHYSICAL column name are reserved: a star-projection exports the
+        physical names, and the two differ whenever a column renames its source
+        (``Column(name="value", sql="_val_0")``). A non-bare ``Column.sql`` is an
+        expression, not a star-exported column, so it contributes nothing.
+
+        Columns that exist in the database but not on the model are outside what
+        SLayer can see without reflection; a physical column literally named
+        ``_val_<n>`` is the only way to hit that residual, which the underscore
+        prefix makes vanishingly unlikely.
+        """
+        names: List[str] = []
+        for c in model.columns:
+            names.append(c.name)
+            sql = getattr(c, "sql", None)
+            if sql and _BARE_IDENT_RE.fullmatch(sql.strip()):
+                names.append(sql.strip())
+        allocator.reserve(*names)
 
     @staticmethod
     def _maybe_quote_ident(ident: Optional[exp.Expression]) -> None:
@@ -5104,7 +5136,7 @@ class SQLGenerator:
         # and never collapse onto one ``_val``; bare refs are type-agnostic
         # (no CAST) and sharing IS correct.
         allocator = self._gen_allocator or self._new_allocator()
-        allocator.reserve(*[c.name for c in source_model.columns])
+        self._reserve_model_column_names(allocator, source_model)
         value_alias_by_sql: Dict[str, str] = {}
 
         def _materialize_if_crossing(
@@ -6673,7 +6705,7 @@ class SQLGenerator:
         # that name — mirrors the host-path reservation in
         # ``_build_first_last_base_select``.
         cte_allocator = self._gen_allocator or self._new_allocator()
-        cte_allocator.reserve(*[c.name for c in target_model.columns])
+        self._reserve_model_column_names(cte_allocator, target_model)
         is_first_or_last = agg_slot.key.agg in ("first", "last")
         grain_extra_projections: List[Tuple[str, exp.Expression]] = []
         for sid in plan.shared_grain_slots:
